refactor(locksmith): Remove unused/deprecated lock endpoints
