--- locksmith/src/types.d.ts
@@ -1,1 +1,0 @@
-declare module '@unlock-protocol/unlock-js'
